+++ src/library/DIPS.Mobile.UI/Components/TextFields/Entry/Entry.cs
@@ -1,3 +1,4 @@
+using DIPS.Mobile.UI.Components.BottomSheets;
 using Colors = DIPS.Mobile.UI.Resources.Colors.Colors;
 
 namespace DIPS.Mobile.UI.Components.TextFields.Entry;